+++ src/main/java/com/project/vetdata/dto/HealthRecordCreateDTO.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 
 public class HealthRecordCreateDTO {
 
-    private Double age;
+    private Integer age;
     private String codPatient;
     private String color;
     private Boolean death;
@@ -27,7 +27,7 @@ public class HealthRecordCreateDTO {
     public HealthRecordCreateDTO() {
     }
 
-    public HealthRecordCreateDTO(Double age, String codPatient, String color, Boolean death,
+    public HealthRecordCreateDTO(Integer age, String codPatient, String color, Boolean death,
                                  Euthanasia euthanasia, Gender gender, String patient, DogSize size,
                                  String tutor, Double weight, Long breedId) {
         this.age = age;
@@ -44,7 +44,7 @@ public class HealthRecordCreateDTO {
     }
 
     @Min(value = 0, message = "Campo deve ser maior que 0")
-    public Double getAge() {
+    public Integer getAge() {
         return age;
     }
 
@@ -95,7 +95,7 @@ public class HealthRecordCreateDTO {
         return breedId;
     }
 
-    public void setAge(Double age) {
+    public void setAge(Integer age) {
         this.age = age;
     }
